+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed 
+
+- **GM-272** - [PLAN] Adjust in share icons
+
 ## [0.17.1] - 2022-09-15
 
 ### Fixed